requirements.txt
Flask
requests
beautifulsoup4
gspread
google-auth
pandas